Minor tests cleanup (#18811)
