+++ tests/framework/web/Post.php
@@ -9,14 +9,30 @@ namespace yiiunit\framework\web;
 
 use yii\base\BaseObject;
 
+/**
+ * @inheritdoc
+ */
 class Post extends BaseObject
 {
+    /**
+     * @var int
+     */
     public $id;
+    /**
+     * @var string
+     */
     public $title;
 
-    public function __construct($id, $title)
+    /**
+     * @param int $id
+     * @param string $title
+     * @param array $config
+     */
+    public function __construct($id, $title, $config = [])
     {
         $this->id = $id;
         $this->title = $title;
+
+        parent::__construct($config);
     }
 }
